producer/requirements.txt
kafka-python==2.0.2
requests
python-telegram-bot==20.7